update rag examples
