+++ examples/basics/rag/adjust-chunk-delimiters/index.ts
@@ -1,6 +1,6 @@
 import { MDocument } from '@mastra/rag';
 
-const doc = MDocument.fromText('Your plain text content...');
+const doc = MDocument.fromText(`Your plain text content...`);
 
 const chunks = await doc.chunk({
   separator: '\n',
