+++ examples/basics/rag/adjust-chunk-size/.gitignore
@@ -1,0 +1,1 @@
+.env
