+++ examples/basics/rag/adjust-chunk-size/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "adjust-chunk-size",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  }
+}
