--- examples/basics/rag/chunk-html.ts
@@ -1,7 +1,0 @@
-import { MDocument } from '@mastra/rag';
-
-const doc = MDocument.fromHTML('<p>Your HTML content...</p>');
-
-const chunks = await doc.chunk();
-
-console.log(chunks);
