+++ examples/basics/rag/chunk-json/.gitignore
@@ -1,0 +1,1 @@
+.env
