+++ examples/basics/rag/chunk-json/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "chunk-json",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  }
+}
