+++ examples/basics/rag/chunk-markdown/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "chunk-markdown",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  }
+}
