+++ examples/basics/rag/chunk-text/.gitignore
@@ -1,0 +1,1 @@
+.env
