+++ examples/basics/rag/chunk-text/package.json
@@ -1,0 +1,10 @@
+{
+  "name": "chunk-text",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  }
+}
