+++ examples/basics/rag/embed-chunk-array/.env.example
@@ -1,0 +1,1 @@
+OPENAI_API_KEY=
