+++ examples/basics/rag/embed-chunk-array/.gitignore
@@ -1,0 +1,1 @@
+.env
