+++ examples/basics/rag/embed-chunk-array/README.md
@@ -1,0 +1,42 @@
+# Embed chunk array
+
+A simple example showing how to embed chunk array using Mastra and OpenAI.
+
+## Prerequisites
+
+- Node.js v20.0+
+- pnpm (recommended) or npm
+- OpenAI API key
+
+## Getting Started
+
+1. Clone the repository and navigate to the project directory:
+
+   ```bash
+   git clone https://github.com/mastra-ai/mastra
+   cd examples/basics/rag/embed-chunk-array
+   ```
+
+2. Copy the environment variables file and add your OpenAI API key:
+
+   ```bash
+   cp .env.example .env
+   ```
+
+   Then edit `.env` and add your OpenAI API key:
+
+   ```env
+   OPENAI_API_KEY=sk-your-api-key-here
+   ```
+
+3. Install dependencies:
+
+   ```
+   pnpm install
+   ```
+
+4. Run the example:
+
+   ```bash
+   pnpm start
+   ```
