+++ examples/basics/rag/embed-chunk-array/index.ts
@@ -1,4 +1,7 @@
 import { MDocument, embed } from '@mastra/rag';
+import dotenv from 'dotenv';
+
+dotenv.config();
 
 const doc = MDocument.fromText('Your text content...');
