+++ examples/basics/rag/embed-chunk-array/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "embed-chunk-array",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.7"
+  }
+}
