+++ examples/basics/rag/embed-text-chunk/.gitignore
@@ -1,0 +1,1 @@
+.env
