+++ examples/basics/rag/embed-text-chunk/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "embed-text-chunk",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.7"
+  }
+}
