+++ examples/basics/rag/embed-text-with-cohere/.env.example
@@ -1,0 +1,1 @@
+COHERE_API_KEY=
