+++ examples/basics/rag/embed-text-with-cohere/.gitignore
@@ -1,0 +1,1 @@
+.env
