+++ examples/basics/rag/embed-text-with-cohere/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "embed-text-with-cohere",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.7"
+  }
+}
