+++ examples/basics/rag/insert-embedding-in-pgvector/.env.example
@@ -1,0 +1,1 @@
+OPENAI_API_KEY=
