+++ examples/basics/rag/insert-embedding-in-pgvector/.gitignore
@@ -1,0 +1,1 @@
+.env
