+++ examples/basics/rag/insert-embedding-in-pinecone/.env.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY=
+PINECONE_API_KEY=
