+++ examples/basics/rag/insert-embedding-in-pinecone/.gitignore
@@ -1,0 +1,1 @@
+.env
