+++ examples/basics/rag/insert-embedding-in-pinecone/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "insert-embedding-in-pinecone",
+  "type": "module",
+  "scripts": {
+    "start": "pnpx tsx index.ts"
+  },
+  "dependencies": {
+    "@mastra/rag": "workspace:*"
+  },
+  "devDependencies": {
+    "dotenv": "^16.4.7"
+  }
+}
