+++ examples/basics/rag/retrieve-results/.env.example
@@ -1,0 +1,2 @@
+OPENAI_API_KEY=
+PINECONE_API_KEY=
