+++ examples/basics/rag/retrieve-results/.gitignore
@@ -1,0 +1,1 @@
+.env
